player_data, daycare, use_item_on_mon, item_data.json

+++ arm9/src/scrcmd_18_c.c
@@ -4,6 +4,7 @@
 #include "save_block_2.h"
 #include "unk_0204639C.h"
 #include "map_header.h"
+#include "scrcmd.h"
 
 extern u16 VarGet(struct UnkStruct_0204639C* arg, u16 wk);
 extern u16 *GetVarPointer(struct UnkStruct_0204639C* arg, u16);
@@ -18,6 +19,6 @@ THUMB_FUNC BOOL ScrCmd_givemon(struct ScriptContext* ctx)
     u16 item = VarGet(ctx->unk80, ScriptReadHalfword(ctx));
     u16 * varPtr = GetVarPointer(ctx->unk80, ScriptReadHalfword(ctx));
     struct PlayerParty * party = SavArray_PlayerParty_get((struct SaveBlock2 *) savePtr->unkC);
-    *varPtr = GiveMon(11, (struct SaveBlock2 *) savePtr->unkC, species, level, item, mapSec, 12);
+    *varPtr = (u16)GiveMon(11, (struct SaveBlock2 *) savePtr->unkC, species, (u8)level, item, mapSec, 12);
     return FALSE;
 }
